Fix possible loss of data warning in RGBController_NZXTHue2.cpp

+++ Controllers/NZXTHue2Controller/RGBController_NZXTHue2.cpp
@@ -409,7 +409,7 @@ void RGBController_NZXTHue2::DeviceUpdateMode()
                     modes[active_mode].speed,
                     direction,
                     colors,
-                    modes[active_mode].colors.size()
+                    (unsigned int)modes[active_mode].colors.size()
                     );
         }
     }
